+++ run/encoder.go
@@ -1,120 +1,17 @@
 package run
 
 import (
-	"bufio"
 	"bytes"
 	"compress/gzip"
 	"context"
 	"errors"
-	"flag"
 	"io"
-	"log"
-	"os"
 	"reflect"
 	"strings"
 
-	"github.com/itzg/go-flagsfiller"
 	"github.com/nextmv-io/sdk"
-	"github.com/nextmv-io/sdk/run/decode"
 	"github.com/nextmv-io/sdk/run/encode"
 )
-
-// GenericDecoder is a Decoder that decodes a json into a struct.
-func GenericDecoder[Input any, Decoder decode.Decoder](
-	_ context.Context, reader any) (input Input, err error,
-) {
-	ioReader, ok := reader.(io.Reader)
-	if !ok {
-		return input, errors.New(
-			"JsonDecoder is not compatible with configured IOProducer",
-		)
-	}
-
-	// Convert to buffered reader and read magic bytes
-	bufferedReader := bufio.NewReader(ioReader)
-	testBytes, err := bufferedReader.Peek(2)
-
-	// Test for gzip magic bytes and use corresponding reader, if given
-	if err == nil && testBytes[0] == 31 && testBytes[1] == 139 {
-		var gzipReader *gzip.Reader
-		if gzipReader, err = gzip.NewReader(bufferedReader); err != nil {
-			return input, err
-		}
-		ioReader = gzipReader
-	} else {
-		// Default case: assume text input
-		ioReader = bufferedReader
-	}
-
-	decoder := *new(Decoder)
-	err = decoder.Decode(ioReader, &input)
-	return input, err
-}
-
-// NoopOptionsDecoder is a Decoder that returns the option as is.
-func NoopOptionsDecoder[Input any](
-	_ context.Context, _ any, input Input,
-) (Input, error) {
-	return input, nil
-}
-
-// FlagParser parses flags and env vars.
-func FlagParser[Option, RunnerCfg any]() (
-	runnerConfig RunnerCfg, option Option, err error,
-) {
-	// create a FlagSetFiller
-	filler := flagsfiller.New(
-		flagsfiller.WithEnv(""),
-		flagsfiller.WithFieldRenamer(
-			func(name string) string {
-				repl := strings.ReplaceAll(name, "-", ".")
-				return strings.ToLower(repl)
-			},
-		),
-	)
-	err = filler.Fill(flag.CommandLine, &option)
-	if err != nil {
-		return runnerConfig, option, err
-	}
-
-	err = filler.Fill(flag.CommandLine, &runnerConfig)
-	if err != nil {
-		return runnerConfig, option, err
-	}
-
-	flag.Parse()
-
-	return runnerConfig, option, nil
-}
-
-// CliIOProducer is a test IOProducer.
-func CliIOProducer(_ context.Context, config any) IOData {
-	cfg, ok := config.(CliRunnerConfig)
-	if !ok {
-		log.Fatal("DefaultIOProducer is not compatible with the runner")
-	}
-	reader := os.Stdin
-	if cfg.Runner.Input.Path != "" {
-		r, err := os.Open(cfg.Runner.Input.Path)
-		if err != nil {
-			log.Fatal(err)
-		}
-		reader = r
-	}
-	var writer io.Writer = os.Stdout
-	if cfg.Runner.Output.Path != "" {
-		w, err := os.Create(cfg.Runner.Output.Path)
-		if err != nil {
-			log.Fatal(err)
-		}
-		writer = w
-	}
-	return NewIOData(
-		reader,
-		nil,
-		writer,
-	)
-}
 
 type version struct {
 	Sdk string `json:"sdk"`
