+++ run/flagparser.go
@@ -1,0 +1,37 @@
+package run
+
+import (
+	"flag"
+	"strings"
+
+	"github.com/itzg/go-flagsfiller"
+)
+
+// FlagParser parses flags and env vars.
+func FlagParser[Option, RunnerCfg any]() (
+	runnerConfig RunnerCfg, option Option, err error,
+) {
+	// create a FlagSetFiller
+	filler := flagsfiller.New(
+		flagsfiller.WithEnv(""),
+		flagsfiller.WithFieldRenamer(
+			func(name string) string {
+				repl := strings.ReplaceAll(name, "-", ".")
+				return strings.ToLower(repl)
+			},
+		),
+	)
+	err = filler.Fill(flag.CommandLine, &option)
+	if err != nil {
+		return runnerConfig, option, err
+	}
+
+	err = filler.Fill(flag.CommandLine, &runnerConfig)
+	if err != nil {
+		return runnerConfig, option, err
+	}
+
+	flag.Parse()
+
+	return runnerConfig, option, nil
+}
